+++ core/i18n/__init__.py
@@ -1,21 +1,17 @@
 from typing import Any
 
 from core.i18n.getBasicInfo import getBasicInfo
-from core.i18n.loadJson import lang, getLangJson
+from core.i18n.getLangJson import lang, getLangJson
 from core.i18n.setLanguage import setLanguage
 
 __all__ = [
     "getBasicInfo",
     "getLangJson",
-    "resetBasicInfo",
-    "basicInfo",
+    "baseInfo",
     "setLanguage",
     "lang",
 ]
 
-basicInfo: dict[Any, Any] = getBasicInfo()
 
-
-def resetBasicInfo() -> None:
-    global basicInfo
-    basicInfo = getBasicInfo() | getLangJson("BasicInfo")
+def baseInfo() -> dict[str, Any]:
+    return getLangJson("BaseInfo")