+++ utils/const.py
@@ -1,3 +1,3 @@
 apiVersion: tuple[int, int, int] = (1, 0, 2)
-sinoteVersion: str = "sinote-2025.1.00862-initial-preview-beta-v0.06.26014"
-__version__: tuple[str, str, str] = ("0", "06", "26012")
+sinoteVersion: str = "sinote-2025.1.00862-initial-preview-beta-v0.06.26015"
+__version__: tuple[str, str, str] = ("0", "06", "26015")